requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pyserial==3.5
requests==2.31.0
matplotlib==3.7.2
